requirements.txt
asgiref==3.6.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
Django==4.1.5
django-appconf==1.0.5
django-bootstrap5==22.2
django-imagekit==4.1.0
djangorestframework==3.14.0
pilkit==2.0
Pillow==9.4.0
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
wincertstore==0.2
